move simulate method in simulate.rs module
